+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Arrays must not exactly match. Right order is enough.
+
 ## [1.0.3] - 2025-12-05
 
 ### Changed
@@ -27,6 +33,7 @@
 
 - Fix pub.dev issues
 
+[Unreleased]: https://github.com/ggsuite/gg_test_matchers/compare/1.0.3...HEAD
 [1.0.3]: https://github.com/ggsuite/gg_test_matchers/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/ggsuite/gg_test_matchers/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/ggsuite/gg_test_matchers/tag/%tag